refactor: top navigation 컴포넌트를 header 디렉토리로 그룹화, HeaderContainer 컴포넌트 생성하여 헤더 스타일 관리 분리 (#123)
