+++ apps/web/components/layout/index.ts
@@ -1,3 +1,3 @@
-export { TopNavigation } from './TopNavigation';
+export { TopNavigation } from './header';
 export { default as BottomNavigation } from './BottomNavigation';
 export { default as PageContainer } from './PageContainer';
